renaming paths for new hdf5writer
